feature-1 Extend MATSim with a utility to generate link flow snapshots
wrote integration code as a test.
wrote an abstract test type which has built-in temp file management.
changed NetworkStateCollector so that it again supports an update() method.
added a toFile() method to it's corresponding object.
some minor changes.
about to run a side-effect test suite and want to save progress before i do - then, i will fix the warnings and errors on this branch.