bug/17-23-24-25 - incident to the integration story - about to refactor population xml file formats.
moving from plans_v4 to population_v6 which supports link-based routes.
plan_v4 is default in matsim example files but only supports node-based paths, and doesn't treat them as a "path" but a required stop on a trip (so, not a strict upper-bound on visited nodes, but just a list of required steps).